+++ wiki/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## 2026-08-01
+
+- [2.25.0] Add `checkTheme()` and `checkIconset()` moved from `sumba`
+- [2.25.0] Add `normalizeMenuItems()` to filter menu items based on route availability
+- [2.25.0] Add `noMenu` option to `wmpa.js` to exclude it from menu items
+
 ## 2026-07-29
 
 - [2.24.4] Bug fix in `wmpa.js`